Blog version 2.0

+++ public/scripts/scripts.js
@@ -820,35 +820,45 @@ $('.btn-add-blog').click(function(){
 });
 
 $('.btn-delete-blog').click(function(){
-      var id = $(this).attr('data-id');
-      var str = $.trim(generateString(20));
-    $.ajax({
-        url: 'deleteBlogAjax',
-        type: "DELETE",
-        data: { id: id },
-		dataType 	: 'json',
-		encode      : true
-    }).done(function(data){
-    	if(data.status == 'deleted') {
 
-    		$("tr[data-id='"+id+"']").fadeOut(function(){
-				$(this).remove();
-			});
+	if(confirm('Are you sure you want to delete this blog?')){
+	      		var id = $(this).attr('data-id');
+	      		var str = $.trim(generateString(20));
+	    	$.ajax({
+		        
+		        url: 'deleteBlogAjax',
+		        type: "DELETE",
+		        data: { id: id },
+				dataType 	: 'json',
+				encode      : true
+	 	
+	 	   }).done(function(data){
+	    	
+	    	if(data.status == 'deleted') {
 
-    		$("tr[data-id='"+id+"']").append('<td><center><span class="alert alert-success" id="'+str+'" style="color:green">A Blog Post has been deleted</span></center><br></td>');
-				$("#"+str).fadeOut(2000, "linear", function(){
-                	$(this).remove();
-        		});	
-    	
-    	} else {
-    		$('#addblog').prepend('<center><span class="alert alert-error" id="'+str+'" style="font-size:20px;color:red">We are having a problem saving your blog</span></center><br>');
-    	}
-    });
+	    		$("tr[data-id='"+id+"']").fadeOut(function(){
+					$(this).remove();
+				});
+
+	    		$("tr[data-id='"+id+"']").append('<td><center><span class="alert alert-success" id="'+str+'" style="color:green">A Blog Post has been deleted</span></center><br></td>');
+					$("#"+str).fadeOut(2000, "linear", function(){
+	                	$(this).remove();
+	        		});	
+	    	
+	    	} else {
+	    		$('#addblog').prepend('<center><span class="alert alert-error" id="'+str+'" style="font-size:20px;color:red">We are having a problem saving your blog</span></center><br>');
+	    	}
+	    
+	    });
+	}
 });
 
 function deleteBlog(obj) {
+if(confirm('Are you sure you want to delete this blog?')){
 	var id = obj.attr('data-id');
-      var str = $.trim(generateString(20));
+    var str = $.trim(generateString(20));
+   
+
     $.ajax({
         url: 'deleteBlogAjax',
         type: "DELETE",
@@ -872,9 +882,14 @@ function deleteBlog(obj) {
     	}
 	});
 }
+}
 
 function editBlog(obj) {
     var id = obj.attr("data-id");
+    $("tr[data-id='"+id+"']").fadeOut(function(){
+				obj.remove();
+	});
+
     $.ajax({
         url: "/blogAjax",
         type: "get",
@@ -890,7 +905,9 @@ function editBlog(obj) {
         $('#googleedit').val(data['google']);
         $('#twitteredit').val(data['twitter']);
         $('#linkedinedit').val(data['linkedin']);
-        $('').attr('src', data['blogheaderimage']);
+        $('#img-render-blogheaderimageedit').attr('src', data['blogheaderimage']);
+        $('#cancel-blog-edit').attr('data-id', data['id']);
+        $('#cancel-blog-edit').attr('data-title', data['title']);
 
         $(document).ready(
             function()
@@ -909,7 +926,6 @@ function editBlog(obj) {
                 });
             });
 
-                console.log(data['body']);
     });
 }
 
@@ -1127,6 +1143,94 @@ $(document).ready(function(){
 
 });
 
-$('#cancel-add-blog').click(function(){
+function cancelBlog() {
+
 	$('#addblog').fadeOut(500, 'linear');
+	$('#editblog').fadeOut(500, 'linear');
+	}
+
+$('#cancel-blog-edit').click(function(){
+	var id = $('#cancel-blog-edit').attr('data-id');
+	var title = $('#cancel-blog-edit').attr('data-title');
+	$("#table-categories tbody").prepend('<tr data-id="'+id+'"><td><span class="category-name">'+title+'</span></td><td><a href="javascript:void(0)" class="bs-btn btn-info btn-edit-category" onclick="editBlog($(this))" data-id="'+id+'">Edit</a> <a href="javascript:void(0)" data-id="'+id+'" onclick="deleteBlog($(this))" class="bs-btn btn-danger btn-delete-blog" data-id="'+id+'">Delete</a></td></tr>');
+});
+
+$(document).ready(function() {
+
+	// process the form
+	$('#editBlogForm').submit(function(event) {
+		var id = $('#cancel-blog-edit').attr('data-id');
+          var formData = new FormData($('#editBlogForm')[0]);
+          var token = $('#addBlogForm > input[name="_token"]').val();
+          formData.append('_token',token);
+          formData.append('id', id);
+
+          var str = $.trim(generateString(20));
+          
+               $.ajax({
+                      url: '/updateBlogAjax',  //Server script to process data
+                      type: 'post',
+                      xhr: function() {  // Custom XMLHttpRequest
+                          var myXhr = $.ajaxSettings.xhr();
+                          if(myXhr.upload){ // Check if upload property exists
+                              myXhr.upload.addEventListener('progress',progressHandler, false); // For handling the progress of the upload
+                          }
+                          return myXhr;
+                      },
+                      //Ajax events
+                      // beforeSend: beforeSendHandler,
+                      success: completeHandler,
+                      error: errorHandler,
+
+                      // Form data
+                      data: formData,
+                      
+                      //Options to tell jQuery not to process data or worry about content-type.
+                      cache: false,
+                      contentType: false,
+                      processData: false,
+                      dataType 	: 'json', // what type of data do we expect back from the server
+                  		encode	: true
+                  });
+            
+          return false;
+
+          function completeHandler(e)
+          {
+             if(e.status == 'saved'){
+             	$('#editblog').fadeOut(500, 'linear');
+				$('#editblog').prepend('<center><span class="alert alert-success" id="'+str+'" style="font-size:20px;color:green">New Blog Post has been added</span></center><br>');
+				$("#"+str).fadeOut(2000, "linear", function(){
+                	$(this).remove();
+        		});	
+
+        		$("#table-categories tbody").prepend('<tr data-id="'+e.id+'"><td><span class="category-name">'+e.title+'</span></td><td><a href="javascript:void(0)" class="bs-btn btn-info btn-edit-category" onclick="editBlog($(this))" data-id="'+e.id+'">Edit</a> <a href="javascript:void(0)" data-id="'+e.id+'" onclick="deleteBlog($(this))" class="bs-btn btn-danger btn-delete-blog" data-id="'+e.id+'">Delete</a></td></tr>');
+
+        		$('#titleedit').val('');
+				$('#facebookedit').val('');
+				$('#googleedit').val('');
+				$('#twitteredit').val('');
+				$('#linkedinedit').val('');
+				$('#img-render-blogheaderimageedit').attr('src', '');
+				$('#btn-editblog-settings-logo').val('');
+
+			} else {
+				$('#editblog').prepend('<center><span class="alert alert-error" id="'+str+'" style="font-size:20px;color:red">We are having a problem saving your blog</span></center><br>');
+			}
+          }
+           function beforeSendHandler()
+          {
+
+          }
+         
+          function progressHandler(e){
+              if(e.lengthComputable){
+                 $('.meter').width(Math.floor(e.loaded/e.total*100) + '%');
+              }
+          }  
+          function errorHandler(e){
+          	console.log(e);
+          }   
+	});
+
 });
